move tests to a dedicated file; satisfy the no-duplicate-conditional-property-access lint
